[refactor] update dependency licenses and remove the aliyun sms dependency
