--- common/src/main/java/org/apache/hertzbeat/common/util/AliYunSendSmsUtil.java
@@ -1,56 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.apache.hertzbeat.common.util;
-
-import com.aliyun.dysmsapi20170525.Client;
-import com.aliyun.dysmsapi20170525.models.SendSmsRequest;
-import com.aliyun.dysmsapi20170525.models.SendSmsResponse;
-import com.aliyun.teaopenapi.models.Config;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import java.util.Map;
-
-/**
- * Alibaba cloud send SMS util
- * Refer from <a href="https://github.com/alibaba/aliyun-spring-boot/blob/master/aliyun-spring-boot-docs/asciidoc/sms.adoc">alibaba aliyun-spring-boot</a>
- */
-public final class AliYunSendSmsUtil {
-
-    private AliYunSendSmsUtil() {
-    }
-
-    private static Client createClient(String accessKeyId, String accessKeySecret) throws Exception {
-        Config config = new Config();
-        config.accessKeyId = accessKeyId;
-        config.accessKeySecret = accessKeySecret;
-        return new Client(config);
-    }
-
-    /**
-     * Method for sending SMS messages: Enter the map format
-     */
-    public static SendSmsResponse send(Map<String, Object> map, String singName, String templateCode, String phone,
-                                       String accessKeyId, String accessKeySecret) throws Exception {
-        Client client = createClient(accessKeyId, accessKeySecret);
-        SendSmsRequest sendReq = new SendSmsRequest()
-                .setPhoneNumbers(phone)
-                .setSignName(singName)
-                .setTemplateCode(templateCode)
-                .setTemplateParam(new ObjectMapper().writeValueAsString(map));
-        return client.sendSms(sendReq);
-    }
-}
